restructure

+++ components/projects.py
@@ -1,19 +1,23 @@
-import imgui
 
+from __future__ import annotations
 import os, glob, json,sys
-
+from copy import deepcopy
+from components.data import BBox, ImageInfo, LabelInfo, Labels
 sys.path.append("./")  # add ROOT to PATH
 import custom_utils
 
 class Project(object):
+    
+    page_status = {}
 
     def __init__(self, name, info_obj) -> None:
         self.name = name
         self.info_obj = info_obj
         self.paths = info_obj["paths"]
         self.model_path = info_obj["model_path"]
-        self.labels = info_obj["labels"]
-        self.imgs = {}
+        self.labels_obj = info_obj["labels"]
+        self.imgs : dict[str,ImageInfo] = {}
+        self.labels : Labels = self.load_labels()
 
     def __str__(self) -> str:
         return json.dumps(self.info_obj, indent=1)
@@ -22,28 +26,74 @@ class Project(object):
         for p in self.paths:
             data_path = p["data"]
             ext_ann_path = p["ext_annotations"]
-            self.import_annotations(data_path, ext_ann_path)
+            # init imgs dict with imgs in data_path
+            for f in glob.glob(data_path + "/*.jpg"):
+                name = os.path.basename(f).rsplit('.')[0]
+                self.imgs[name] = ImageInfo(name, f)
+        self.load_annotations()        
         
-    def import_annotations(self, data_path, yolo_annotations_path):
+    def import_yolo_annotations(self, data_path, yolo_annotations_path):
 
         for f in glob.glob(f"{yolo_annotations_path}/*.txt"):
             name = os.path.basename(f).rsplit('.')[0]
+
             img_path = f"{data_path}/{name}.jpg"
-            img_size = custom_utils.get_image_size(img_path)
-            bboxes = custom_utils.import_yolo_predictions(f, img_size[0], img_size[1])
-            self.imgs[img_path] = {}
-            with open(f"projects/{self.name}/annotations/{name}.json", "w") as fp:
-                json.dump(bboxes, fp, indent=1)
+            img_info = ImageInfo(name, img_path)
+            bboxes = custom_utils.import_yolo_predictions(f, img_info.w, img_info.h)
 
-    def get_pair(self):
+            img_info.add_bboxes(bboxes)
+            self.imgs[name] = img_info
+            
+            if not os.path.exists(f"projects/{self.name}/annotations/{name}.json"):
+                bboxes_list = list(map(lambda x: x.as_obj(), bboxes))
+                with open(f"projects/{self.name}/annotations/{name}.json", "w") as fp:
+                    data = {"img_path": img_path, "bboxes": bboxes_list}
+                    json.dump(data, fp, indent=1)
+    
+    def load_annotations(self, ):
+
+        for f in glob.glob(f"projects/{self.name}/annotations/*.json"):
+            name = os.path.basename(f).rsplit('.')[0]
+            with open(f, "r") as fp:
+                data = json.load(fp)
+
+            bboxes = list(map(lambda x: BBox(x["xmin"],x["ymin"],x["xmax"],x["ymax"], x["label"], x["conf"]), data["bboxes"]))
+            img_info = self.imgs[name]
+            
+            img_info.add_bboxes(bboxes)
+            self.imgs[name] = img_info
+        
+
+    def get_image(self):
         for img in self.imgs:
-            if self.imgs[img].get("size") is None:
-                self.imgs[img]["orig_size"] = custom_utils.get_image_size(img)
-            self.imgs[img]["name"] = os.path.basename(img).rsplit('.')[0]
-            if self.imgs[img].get("bboxes") is None:
-                self.imgs[img]["bboxes"] = self.load_json_annotation(self.imgs[img]["name"])
-
             yield self.imgs[img]
+
+    def load_labels(self):
+
+        labels = Labels()
+        for k in self.info_obj["labels"]:
+            obj = self.info_obj["labels"][k]
+            l = LabelInfo(
+                obj["index"], 
+                obj["label"],
+                [ color / 255.0 for color in obj["rgb"]],
+                obj["shortcut"])
+            labels.add_label(l)
+        return labels
+    
+    def update_labels(self, save=True):
+
+        labels_obj = {}
+        for o in self.labels:
+            labels_obj[o.index] = o.as_obj(rgb_int=True)
+        
+        self.info_obj["labels"] = labels_obj
+        if save:
+            self.save_config()
+
+    def save_config(self):
+        with open(f"projects/{self.name}/info.json", "w") as fp:
+            json.dump(self.info_obj, fp, indent=1)
 
     def load_json_annotation(self, img_name):
 
@@ -51,13 +101,30 @@ class Project(object):
             annotations = json.load(fp)
         
         return annotations
+    
+    def set_changed(self, img_name):
+
+        self.page_status[img_name] = True
+
+    def save_annotations(self,):
+        
+        for img_info in self.imgs.values():
+            
+            if img_info.is_changed:
+                with open(f"projects/{self.name}/annotations/{img_info.name}.json", "w") as fp:
+                    scaled_bboxes = []
+                    for bbox in img_info.bboxes:
+                        scaled_bboxes.append(bbox.scale((img_info.w, img_info.h), (img_info.orig_w, img_info.orig_h)).as_obj())
+                    data = {"img_path": img_info.path, "bboxes": scaled_bboxes}
+                    json.dump(data, fp, indent=1 )
+                img_info.set_changed(False)
 
 
-def get_projects():
+def load_projects():
 
     projects = []
     for p in glob.glob("projects/*"):
-        # print(os.path.basename(p))
+        
         with open(p + "/info.json", "r") as fp:
             info_obj = json.load(fp)
         projects.append(Project(os.path.basename(p), info_obj))
@@ -65,8 +132,6 @@ def get_projects():
     return projects
 
 
-
-
 if __name__ == "__main__":
-    for p in get_projects():
+    for p in load_projects():
         print(p)
